+++ web/src/modules/travel-authorizations/pages/travel-authorization-edit/DetailsTab.vue
@@ -24,7 +24,7 @@
       <v-btn
         class="ml-3"
         color="secondary"
-        :to="{ name: 'TravelFormList' }"
+        :to="{ name: 'MyTravelAuthorizationsPage' }"
         >Back</v-btn
       >
     </div>
